requirements.txt

flask
python-dotenv
yagmail